% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GiniEstimation.R
\name{GiniEstimation}
\alias{GiniEstimation}
\title{GiniEstimation}
\usage{
GiniEstimation(cdf, mu, n, m, a, b)
}
\arguments{
\item{cdf}{for the supposed distribution(continuous)}

\item{mu}{for theoritical expectation}

\item{n}{for size of random numbers}

\item{m}{for repeating times}

\item{a}{for the better xlim for histogram(u can choose a better one when u observed its distribution in histogram with random a b )}

\item{b}{for the better xlim for histogram(u can choose a better one when u observed its distribution in histogram with random a b )}
}
\value{
mean median and deciles ,a density histograms of the replicates\code{n}
}
\description{
Estimate by simulation the mean, median and deciles of hat{G} if X is continuous random variable
}
\examples{
\dontrun{
GiniEstimation(rlnorm,exp(0.5),1e3,1e2,0.45,0.65)#The standard lognormal
}
}
