% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteofCauchy.R
\name{InteofCauchy}
\alias{InteofCauchy}
\title{InteofCauchy}
\usage{
InteofCauchy(q, eta, theta)
}
\arguments{
\item{q}{for distribution function F(q)}

\item{eta}{the location parameter}

\item{theta}{the scale parameter}
}
\value{
a value \code{n}
}
\description{
a function to compute the cdf of the Cauchy distribution.
}
\examples{
\dontrun{
InteofCauchy(22,0,1)
}
}
