% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCofBeta.R
\name{MCofBeta}
\alias{MCofBeta}
\title{a Monte Carlo estimate of the Beta function}
\usage{
MCofBeta(n, a, b)
}
\arguments{
\item{n}{the number of sampling times}

\item{a}{for Beta function}

\item{b}{for Beta function}
}
\value{
a value \code{n}
}
\description{
a Monte Carlo estimate of the Beta(a, b) cdf
}
\examples{
\dontrun{
MCofBeta(1e4,3,3)
c(MCofBeta(1e4,3,3),beta(3,3))
#compare the MC estimation and the theoretical value
}
}
